task_next_child() s/unsigned int pos/loff_t pos
pos should be loff_t type. It's might be fine to truncate it
to unsigned int in this case, but it's better not to.
E.g. see 9f6e963f06c1 ("proc: fix ->f_pos overflows in first_tid()")

+++ fs/proc/array.c
@@ -651,8 +651,8 @@ int proc_pid_statm(struct seq_file *m, struct pid_namespace *ns,
 }
 
 #ifdef CONFIG_PROC_CHILDREN
-struct task_struct *
-task_next_child(struct task_struct *parent, struct task_struct *prev, unsigned int pos)
+struct task_struct *task_next_child(struct task_struct *parent,
+				struct task_struct *prev, loff_t pos)
 {
 	struct task_struct *task;
